requirements.txt
paho-mqtt>=2.0.0
requests>=2.28.0
beautifulsoup4>=4.10.0
lxml>=4.6.3
